MDEV-20091 DROP TEMPORARY table is logged despite no CREATE was logged
MDEV-5589 commit set up a policy to skip DROP TEMPORARY TABLE
binary logging in case the target table has not been "CREATEed" in
binlog (no CREATE Query-log-event was logged into the binary log).

It turns out that

1. the rule did not cover non-existing table DROPped with IF-EXISTS
   clause. The logged-create knowledge for the non-existing one does
   not even need MDEV-5589 patch, and

2. connection close disobeys it to trigger automatic DROP-IF-EXISTS
   binlogging.

Either 1 or 2 or even both is/are also responsible for unexpected
binlog records observed in MDEV-17863, actually rendering
a referred @@global.read_only irrelevant as far as the described stored
procedure definition *and* the ROW binlog-format are concerned.